a2a red deals part 3/requirements.txt
pydantic==2.6.2
httpx==0.27.0
tiktoken==0.5.2
requests ==2.32.3
tiktoken ==0.5.2
MarkupSafe ==2.1.3
fastapi==0.108.0
python-dotenv==1.0.0
pydantic==2.5.0
uvicorn==0.23.2
datetime==5.1
jsonschema==4.18.0
configparser==5.3.0